feat: implement position, chain, and analytics gRPC handlers (Phase 3)
Add the remaining three Phase 3 gRPC handlers and their proto definitions,
completing all handler work before server binary wiring.

Proto / generated:
- position.proto — PositionService (ListPositions, GetPosition) with
  PositionStatus enum (OPEN/CLOSED/UNSPECIFIED) replacing bool flags
- chain.proto — ChainService (GetChain) returning ChainDetail with
  per-event leg data (action, instrument, lot size); LinkType enum
- analytics.proto — AnalyticsService (GetAccountSummary); symbol/strategy
  analytics deferred to phase6.md

Domain:
- Add LinkTypeOpen, ChainDetail, ChainEvent, ChainEventLeg to domain/chain.go
- Document Transaction.Quantity as always non-negative (direction in Action)

Service:
- ChainService.GetChainDetail: account-scoped, batch-loads transactions,
  deduplicates trade IDs, collapses both not-found cases to ErrNotFound
- New ChainReader and AccountSummaryReader interfaces
- PositionReader.ListPositions gains closedOnly bool (parallel to trades)
- PnLSummary.PositionsClosed typed int32 at source to avoid handler cast

Repository:
- PositionRepository.ListPositions and sqlite impl updated for closedOnly
- ListPositions returns empty slice (not nil) when no rows match

Handlers:
- position_handler: maps PositionStatus enum → openOnly/closedOnly;
  realized_pnl only populated for closed positions
- chain_handler: thin mapping of ChainDetail → proto ChainDetail
- analytics_handler: from/to are required (InvalidArgument if absent)

Tests:
- trade_handler_test.go (previously missed)
- Full handler test coverage for position, chain, analytics

Docs:
- docs/phase6.md: post-MVP analytics (GetSymbolPerformance,
  GetStrategyPerformance)
- docs/future.md: computeStrikeExpChange DST comment note

+++ internal/service/analytics_service_test.go
@@ -73,7 +73,7 @@ func TestAnalyticsService_EmptyRange(t *testing.T) {
 
 	summary, err := asvc.GetPnLSummary(ctx, acc.ID, allTime, now)
 	require.NoError(t, err)
-	assert.Equal(t, 0, summary.PositionsClosed)
+	assert.Equal(t, int32(0), summary.PositionsClosed)
 	assert.True(t, decimal.Zero.Equal(summary.WinRate))
 	assert.True(t, decimal.Zero.Equal(summary.RealizedPnL))
 
@@ -116,7 +116,7 @@ func TestAnalyticsService_WinningPosition(t *testing.T) {
 
 	summary, err := asvc.GetPnLSummary(ctx, acc.ID, allTime, future)
 	require.NoError(t, err)
-	assert.Equal(t, 1, summary.PositionsClosed)
+	assert.Equal(t, int32(1), summary.PositionsClosed)
 	assert.True(t, decimal.NewFromFloat(1).Equal(summary.WinRate), "win rate should be 1.0")
 	assert.True(t, decimal.NewFromInt(300).Equal(summary.RealizedPnL), "got %s", summary.RealizedPnL)
 }
@@ -149,7 +149,7 @@ func TestAnalyticsService_LosingPosition(t *testing.T) {
 
 	summary, err := asvc.GetPnLSummary(ctx, acc.ID, allTime, future)
 	require.NoError(t, err)
-	assert.Equal(t, 1, summary.PositionsClosed)
+	assert.Equal(t, int32(1), summary.PositionsClosed)
 	assert.True(t, decimal.Zero.Equal(summary.WinRate), "win rate should be 0")
 	assert.True(t, summary.RealizedPnL.IsNegative(), "P&L should be negative, got %s", summary.RealizedPnL)
 }
@@ -190,7 +190,7 @@ func TestAnalyticsService_MixedWinRate(t *testing.T) {
 
 	summary, err := asvc.GetPnLSummary(ctx, acc.ID, allTime, future)
 	require.NoError(t, err)
-	assert.Equal(t, 2, summary.PositionsClosed)
+	assert.Equal(t, int32(2), summary.PositionsClosed)
 	half := decimal.NewFromFloat(0.5)
 	assert.True(t, half.Equal(summary.WinRate), "got %s", summary.WinRate)
 	// Net P&L: 300 − 200 = 100
@@ -276,7 +276,7 @@ func TestAnalyticsService_DateRangeFilter(t *testing.T) {
 
 	summary, err := asvc.GetPnLSummary(ctx, acc.ID, from, to)
 	require.NoError(t, err)
-	assert.Equal(t, 1, summary.PositionsClosed, "only the April position should be counted")
+	assert.Equal(t, int32(1), summary.PositionsClosed, "only the April position should be counted")
 	assert.True(t, decimal.NewFromInt(1).Equal(summary.WinRate))
 
 	pnl, err := asvc.GetSymbolPnL(ctx, acc.ID, "SPY", from, to)
